contract_analyzer/contract_analyzer_analysis_ms/requirements.txt
pymongo
fastapi[standard]
pydantic
uvicorn
python-dotenv